Add support for SEO split tests
SEO split tests differ from User split tests in that variant depends on
the path rather than the user. To evenly distribute the variants between
paths we hash the path then take the modulo of the number of variants
variants.get(hash(path) % variants.size()), this only works if there are
a large number of unique paths served by the same endpoint.