Add cookie consent option to experiment filter
If the cookie consent is set but the cookie consent cookie does not
exist or does not match the regex pattern provided then mitosis will not
set any cookies. This means we will be unable to run user experiments on
users who do not accept the cookie policy (seo experiments don't use
cookies).

The code itself is fairly simple but due to the nature of
HttpServletRequest and HttpServletRespons testing the code is relatively
complex, requiring us to mock and stub the request and response objects
in order to check the outcome of applying the filter.