Set request attributes even when cookie consent not accepted
If the ExperimentFilter is configured then the 'experiments' request
attribute should always be a map. The map may be empty if no
experiments are configured or if the cookie consent has not be accepted
by the user. It is better to have an empty map than a null value as it
prevents null pointer exceptions
 - in the code that checks the map for the existance of an experiment.
 - in Spring @RequestParameter arguments
 - in Thymeleaf templates

---

SEO experiments do not rely on cookies but do need the experiment
variants to be set in the 'experiments' request attribute. SEO
experiment variants should be set regardless of cookie consent status.